+++ Changes.md
@@ -1,6 +1,8 @@
 HEAD
 -----------
 
+- Add capistrano support, just require 'sidekiq/capistrano' in config/deploy.rb.
+- Workers now log upon start and finish (mperham)
 - Messages for terminated workers are now automatically requeued (mperham)
 - Add support for Exceptional error reporting (bensie)